feat: move image alignment and size to style
